CONTRIBUTING.md file added
When I started hacking on the project, I did not have
any direction on which branch to contribute to.
Offline I got a direct message that our new convention
is that PRs to this project should use develop as the
base and that unstable should only be used to merge
multiple PRs that have effects on each other.

Let's actually cement this so that it isn't by
word of mouth and that anyone can pick it up.