Add test case for non-incremental nonces
This test case checks that transactions have to have strictly increasing
nonces. It tests by sending a transaction with nonce 0, and then tries
to send a transaction with nonce 2. The second transaction should fail,
since it is not strictly increasing.

Some of the error descriptions have changed, but no one should be
matching for a specific error string.