Design for the future of Ethermint (#295)
This document explains the ideal Ethermint
architecture. The current design is similar
but less modular. From this starting point,
we will start an incremental refactor until
we are at this design. The design itself is
not perfect and will change over time.

By splitting Ethermint up this way, we have
the ability to add custom RPC endpoints.
This allows the addition of special IBC
endpoints which will be needed for IBC.

This adds a description of a light client
for ethermint which leverages the properties
of tendermint to provide a fully secure EVM
light client. A light client can keep of
with the state and doesn't have to trust a
single node. Furthermore, it can submit
transactions.

I've added a paragraph that explains how
accounts are managed by an ethermint instance.
This is the last missing piece as now ethermint
provides all the functionality in a coherent
way that we need it to.

* Extend architecture with IBC and Rewards

* Reason why we need this redesign

* Explanation for accounts

* Add time estimate

* Simplify future design

* More IBC explanations

* Light client explanation