Add stateful checkTx (#276)
* Add stateful checkTx

* Tests for stateful checktx

* Mirror validateTx logic from tx_pool.go. Move checkTxState updates

* Move transaction size limit into const. Remove uneeded fmt import

* Enable nonces to be strictly increasing

* Extract all directory setup into functional style tests

This change removes the duplicated code to setup temporary directories
for each test.
Now all code is grouped in the setupTestCase() function
instead of being littered around all the test cases.
There are no side effects of this.

* Refactor app_test.go to remove the duplicated code

The tests where getting unwieldy since they duplicated a lot of code.
With the new setup it is much easier to create new test cases, since the
setup and tear-down are handled nicely.

There are no side effects.

* Add test case for non-incremental nonces

This test case checks that transactions have to have strictly increasing
nonces. It tests by sending a transaction with nonce 0, and then tries
to send a transaction with nonce 2. The second transaction should fail,
since it is not strictly increasing.

Some of the error descriptions have changed, but no one should be
matching for a specific error string.

* Refactor tests and RPC library

This change is the last refactor of the app_tests and was needed for
clarity.

Also, it upgrades everything to use a typed RPC client instead of
passing strings to get to raw endpoints.

* Add check whether it's contract creation

* Update Changelog

Signed-off-by: Adrian Brink <adrian@brink-holdings.com>